Rename TString.py to strings
