+++ src/rootfilespec/bootstrap/strings.py
